moved parameters functions to separate file
